pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "conda-forge-only"
version = "0.1.0"
description = "A pre-commit hook to enforce only 'conda-forge' in YAML channel lists."
authors = [{ name = "Your Name", email = "your.email@example.com" }]
dependencies = [
    "pyyaml"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
